- using nexmo for smsing - better reply text - improved sss reply

+++ src/shoutit/management/commands/reply_sss.py
@@ -7,6 +7,7 @@ import datetime
 from django.core.management.base import BaseCommand
 from django.utils import timezone
 from shoutit.controllers import message_controller
+from shoutit.controllers.notifications_controller import sms_sss_user
 from shoutit.models import Conversation, DBCLConversation
 import random
 
@@ -24,10 +25,10 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         # get conversations
         now = timezone.now()
-        nine_hours_ago = now + datetime.timedelta(hours=-9)
+        four_hours_ago = now + datetime.timedelta(hours=-0)
         days_ago = now + datetime.timedelta(days=-options['days'])
         conversations = Conversation.objects.filter(created_at__gte=days_ago,
-                                                    created_at__lt=nine_hours_ago,
+                                                    created_at__lt=four_hours_ago,
                                                     shout__is_sss=True)
         replies_count = 0
         for conversation in conversations:
@@ -37,22 +38,32 @@ class Command(BaseCommand):
                 # sss user already replied
                 continue
             reply_sss(conversation, shout, sss_user)
-            # delete dbcl conversations
-            DBCLConversation.objects.filter(to_user=sss_user).delete()
             replies_count += 1
         self.stdout.write("Successfully replied on behalf of %s sss users" % replies_count)
 
 
 def reply_sss(conversation, shout, sss_user):
     # set the text
-    if sss_user.profile.country in ['JO', 'EG', 'SA', 'OM', 'BH']:
-        text = random.choice(arabic_replies)
-    else:  # ['AE', 'QA', 'KQ', 'BH', ...]
-        text = random.choice(english_replies)
-    # send the message
-    message_controller.send_message(conversation=conversation, user=sss_user, text=text)
-    # leave conversation
-    conversation.mark_as_deleted(sss_user)
-    # disable the shout
-    shout.is_disabled = True
-    shout.save()
+    sss_profile = sss_user.profile
+    mobile = sss_profile.mobile
+    if mobile:
+        text = mobile
+        last_message = conversation.last_message
+        # send the message
+        message_controller.send_message(conversation=conversation, user=sss_user, text=text)
+        # sms the sss_user again
+        sms_sss_user(sss_user, from_user=last_message.user, message=last_message, sms_anyway=True)
+    else:
+        if sss_user.profile.country in ['JO', 'EG', 'SA', 'OM', 'BH']:
+            text = random.choice(arabic_replies)
+        else:  # ['AE', 'QA', 'KQ', 'BH', ...]
+            text = random.choice(english_replies)
+        # send the message
+        message_controller.send_message(conversation=conversation, user=sss_user, text=text)
+        # leave conversation
+        conversation.mark_as_deleted(sss_user)
+        # disable the shout
+        shout.is_disabled = True
+        shout.save()
+        # delete dbcl conversations
+        DBCLConversation.objects.filter(to_user=sss_user).delete()
